+++ test/respond-ok-test.js
@@ -1,41 +1,44 @@
 // --------------------------------------------------------------------------
-// Copyright (C) 2016-2022 Nick Mitchell
-// MIT Licensed
+// Copyright (C) 2016-2023 Nick Mitchell - MIT Licensed
 // --------------------------------------------------------------------------
 /* eslint-env mocha */
 /* eslint-disable no-unused-expressions */
 /* eslint-disable no-magic-numbers */
 'use strict';
 
-const express    = require('express');
-const request    = require('supertest');
-const HttpStatus = require('http-status-codes').StatusCodes;
-const expect     = require('chai').expect;
+const express = require('express');
+const request = require('supertest');
+const status  = require('http-status-codes').StatusCodes;
+const expect  = require('chai').expect;
 
-const api = require('../');
+const api    = require('..');
+const helper = require('./helper');
 
 // --------------------------------------------------------------------------
 
 var app = express();
 
-describe('lmc-api-util - makeOk() and children', function() {
-	describe('makeOk()', function() {
+describe('lmc-api-util - respondOk() and children', function() {
+	describe('respondOk()', function() {
 		before(function() {
 			app.get('/ok', function(req, res) {
 				if(req.query.resultobj) {
-					api.makeOk(res, 'OK result with object', {foo: 'bar'});
+					api.respondOk(res, 'OK result with object', {foo: 'bar'});
 				} else if(req.query.nomessage) {
-					api.makeOk(res);
+					api.respondOk(res);
 				} else {
-					api.makeOk(res, 'OK result');
+					api.respondOk(res, 'OK result');
 				}
 			});
+		});
+		it('should throw an error if passed something besides a result object', function() {
+			helper.expectErrorWithBadResultObject(api.respondOk);
 		});
 		it('should return OK result with message', function(done) {
 			request(app)
 				.get('/ok')
 				.set('Accept', 'application/json')
-				.expect(HttpStatus.OK)
+				.expect(status.OK)
 				.end(function(err, res) {
 					expect(err).to.be.null;
 					expect(res.body.result).to.equal('OK');
@@ -47,7 +50,7 @@ describe('lmc-api-util - makeOk() and children', function() {
 			request(app)
 				.get('/ok?nomessage=true')
 				.set('Accept', 'application/json')
-				.expect(HttpStatus.OK)
+				.expect(status.OK)
 				.end(function(err, res) {
 					expect(err).to.be.null;
 					expect(res.body.result).to.equal('OK');
@@ -59,7 +62,7 @@ describe('lmc-api-util - makeOk() and children', function() {
 			request(app)
 				.get('/ok?resultobj=true')
 				.set('Accept', 'application/json')
-				.expect(HttpStatus.OK)
+				.expect(status.OK)
 				.end(function(err, res) {
 					expect(err).to.be.null;
 					expect(res.body.result).to.equal('OK');
@@ -69,13 +72,13 @@ describe('lmc-api-util - makeOk() and children', function() {
 				});
 		});
 	});
-	describe('makeFound()', function() {
+	describe('respondFound()', function() {
 		before(function() {
 			app.get('/found', function(req, res) {
 				if(req.query.type === 'object') {
-					api.makeFound(res, 'Thing 123', {thing: {stuff: 'junk'}});
+					api.respondFound(res, 'Thing 123', {thing: {stuff: 'junk'}});
 				} else if(req.query.type === 'array') {
-					api.makeFound(res, 'Thing 123', {
+					api.respondFound(res, 'Thing 123', {
 						things: [
 							{id: 1},
 							{id: 2},
@@ -83,15 +86,18 @@ describe('lmc-api-util - makeOk() and children', function() {
 						]
 					});
 				} else {
-					api.makeFound(res, 'Thing 123', {thing: 'junk'});
+					api.respondFound(res, 'Thing 123', {thing: 'junk'});
 				}
 			});
+		});
+		it('should throw an error if passed something besides a result object', function() {
+			helper.expectErrorWithBadResultObject(api.respondFound);
 		});
 		it('should return OK with "Found" message and extra params', function(done) {
 			request(app)
 				.get('/found')
 				.set('Accept', 'application/json')
-				.expect(HttpStatus.OK)
+				.expect(status.OK)
 				.end(function(err, res) {
 					expect(err).to.be.null;
 					expect(res.body.result).to.equal('OK');
@@ -104,7 +110,7 @@ describe('lmc-api-util - makeOk() and children', function() {
 			request(app)
 				.get('/found?type=object')
 				.set('Accept', 'application/json')
-				.expect(HttpStatus.OK)
+				.expect(status.OK)
 				.end(function(err, res) {
 					expect(err).to.be.null;
 					expect(res.body.result).to.equal('OK');
@@ -118,7 +124,7 @@ describe('lmc-api-util - makeOk() and children', function() {
 			request(app)
 				.get('/found?type=array')
 				.set('Accept', 'application/json')
-				.expect(HttpStatus.OK)
+				.expect(status.OK)
 				.end(function(err, res) {
 					expect(err).to.be.null;
 					expect(res.body.result).to.equal('OK');
@@ -132,17 +138,20 @@ describe('lmc-api-util - makeOk() and children', function() {
 				});
 		});
 	});
-	describe('makeFoundZero()', function() {
+	describe('respondFoundZero()', function() {
 		before(function() {
 			app.get('/foundzero', function(req, res) {
-				api.makeFoundZero(res, 'Things', {stuff: 'junk'});
+				api.respondFoundZero(res, 'Things', {stuff: 'junk'});
 			});
+		});
+		it('should throw an error if passed something besides a result object', function() {
+			helper.expectErrorWithBadResultObject(api.respondFoundZero);
 		});
 		it('should return OK with "Found zero" message', function(done) {
 			request(app)
 				.get('/foundzero')
 				.set('Accept', 'application/json')
-				.expect(HttpStatus.OK)
+				.expect(status.OK)
 				.end(function(err, res) {
 					expect(err).to.be.null;
 					expect(res.body.result).to.equal('OK');
@@ -152,20 +161,23 @@ describe('lmc-api-util - makeOk() and children', function() {
 				});
 		});
 	});
-	describe('makeCreated()', function() {
+	describe('respondCreated()', function() {
 		before(function() {
 			app.get('/created', function(req, res) {
-				api.makeCreated(res, 'Thing 123', {
+				api.respondCreated(res, 'Thing 123', {
 					location: '/created/123',
 					stuff: 'junk'
 				});
 			});
 		});
+		it('should throw an error if passed something besides a result object', function() {
+			helper.expectErrorWithBadResultObject(api.respondCreated);
+		});
 		it('should return CREATED status with "Created new" message and location header', function(done) {
 			request(app)
 				.get('/created')
 				.set('Accept', 'application/json')
-				.expect(HttpStatus.CREATED)
+				.expect(status.CREATED)
 				.end(function(err, res) {
 					expect(err).to.be.null;
 					expect(res.header.location).to.equal('/created/123');
@@ -177,20 +189,23 @@ describe('lmc-api-util - makeOk() and children', function() {
 				});
 		});
 	});
-	describe('makeUpdated()', function() {
+	describe('respondUpdated()', function() {
 		before(function() {
 			app.get('/updated', function(req, res) {
-				api.makeUpdated(res, 'Thing 123', {
+				api.respondUpdated(res, 'Thing 123', {
 					location: '/updated/123',
 					stuff: 'junk'
 				});
 			});
 		});
+		it('should throw an error if passed something besides a result object', function() {
+			helper.expectErrorWithBadResultObject(api.respondUpdated);
+		});
 		it('should return OK status with "Updated" message and location header', function(done) {
 			request(app)
 				.get('/updated')
 				.set('Accept', 'application/json')
-				.expect(HttpStatus.OK)
+				.expect(status.OK)
 				.end(function(err, res) {
 					expect(err).to.be.null;
 					expect(res.header.location).to.equal('/updated/123');
@@ -202,17 +217,20 @@ describe('lmc-api-util - makeOk() and children', function() {
 				});
 		});
 	});
-	describe('makeDeleted()', function() {
+	describe('respondDeleted()', function() {
 		before(function() {
 			app.get('/deleted', function(req, res) {
-				api.makeDeleted(res, 'Thing 123', {stuff: 'junk'});
+				api.respondDeleted(res, 'Thing 123', {stuff: 'junk'});
 			});
+		});
+		it('should throw an error if passed something besides a result object', function() {
+			helper.expectErrorWithBadResultObject(api.respondDeleted);
 		});
 		it('should return OK with "Deleted" message', function(done) {
 			request(app)
 				.get('/deleted')
 				.set('Accept', 'application/json')
-				.expect(HttpStatus.OK)
+				.expect(status.OK)
 				.end(function(err, res) {
 					expect(err).to.be.null;
 					expect(res.body.result).to.equal('OK');