+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # lmc-api-util Changelog
 
+## 2.0.0
+> Released 16 Jun 2021
+
+### New
+- Added type checking of response object in `respond` functions, throwing an Error if it is not a valid `express` response object
+- Added `validateRequest()` to not only check for required query values but check the type of values
+- Added object functions for paging/filtering/sorting objects
+- Added query functions for generating sequelize queries for paging/filtering/sorting objects
+
+### Updated
+- Renamed `make` series of functions to `respond` to better reflect action of function. Legacy `make` functions are
+  still available for backward compatibility but are undocumented
+- Added support for *JSON Server* style paging parameters (`_page` and `_limit`) in `calcPaging()`
+- Deprecated `checkRequired()` function, as it was superceded by `validateRequest()`. It is still available for backward
+  compatibility but is undocumented
+
 ## 1.3.0
 > Released 14 Dec 2022